Break register form v1 down into sub-components
